style: move Loader
